Refactor method name and unit test update
Change method name for creating list, add unit test for factory and update in travis yml

Issues:
#61

Contributors:
@krasen86
